Move files to more descriptive location
