--- libs/eo/transfer/README.md
@@ -1,1 +1,0 @@
-# Transfer Page
