+++ libs/eo/transfers/src/index.ts
@@ -14,4 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './eo-transfer.module';
+export * from './eo-transfers.module';
